% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_tree-plot.R
\name{create_dot_body}
\alias{create_dot_body}
\title{Writes each node information
If it is a leaf node: show it in different color, show number of samples, show leaf id
If it is a non-leaf node: show its splitting variable and splitting value}
\usage{
create_dot_body(tree, index = 1)
}
\arguments{
\item{tree}{the tree to convert}

\item{index}{the index of the current node}
}
\description{
Writes each node information
If it is a leaf node: show it in different color, show number of samples, show leaf id
If it is a non-leaf node: show its splitting variable and splitting value
}
\keyword{internal}
