% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_tree-plot.R
\name{export_graphviz}
\alias{export_graphviz}
\title{Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into \code{dot_string}.}
\usage{
export_graphviz(tree)
}
\arguments{
\item{tree}{the tree to convert}
}
\description{
Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into \code{dot_string}.
}
\keyword{internal}
