% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_tree.R
\name{find_leaf_node}
\alias{find_leaf_node}
\title{Query a tree with a sample}
\usage{
find_leaf_node(tree, sample)
}
\arguments{
\item{tree}{policy_tree tree object}

\item{sample}{A vector of observations \eqn{X_i}}
}
\value{
The leaf node the sample falls in to.
}
\description{
Query a tree with a sample
}
\keyword{internal}
