% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gen_data_epl}
\alias{gen_data_epl}
\title{Example data generating process from Policy Learning With Observational Data}
\usage{
gen_data_epl(n, type = c("continuous", "jump"))
}
\arguments{
\item{n}{Number of observations}

\item{type}{tau is "continuous" (default - equation 46) or exhibits "jumps" (equation 47)}
}
\value{
A list
}
\description{
The DGP from section 5.2 in Athey and Wager (2020)
}
\references{
Athey, Susan, and Stefan Wager. "Policy Learning With Observational Data." Econometrica 89.1 (2020): 133-161.
}
