% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_test_tree}
\alias{predict_test_tree}
\title{Predict with the above test tree.}
\usage{
predict_test_tree(tree, newdata)
}
\arguments{
\item{tree}{tree from make_tree}

\item{newdata}{data matrix}
}
\value{
Vector of predictions
}
\description{
Predict with the above test tree.
}
\keyword{internal}
