% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_causal_forest-print.R
\name{print.multi_causal_forest}
\alias{print.multi_causal_forest}
\title{Print a multi_causal_forest object.}
\usage{
\method{print}{multi_causal_forest}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a multi_causal_forest object.
}
