% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_tree-print.R
\name{print.policy_tree}
\alias{print.policy_tree}
\title{Print a policy_tree object.}
\usage{
\method{print}{policy_tree}(x, ...)
}
\arguments{
\item{x}{The tree to print.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a policy_tree object.
}
