% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_tree-plot.R
\name{plot.policy_tree}
\alias{plot.policy_tree}
\title{Plot a policy_tree tree object.}
\usage{
\method{plot}{policy_tree}(x, ...)
}
\arguments{
\item{x}{The tree to plot}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a policy_tree tree object.
}
